Create wage.c
This is a wage calculator. the idea behind this is that user input of hours is taken and then multiplied with the pay per hours.
tax is also deducted automatically. if pay is <=300 then 15% is charged then for the next 150$ 20% is charged and then for the remaining 25% is charged. So in this program tax growns progressively. Overtime hours (i.e any no. of hours after 40 ) are also calculated and CTC and Taxes are calculated accordingly.